requirements.txt
speech-recognition-fork == 3.8.1.2021.6.14

